% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results.R
\name{get_intervention_levels}
\alias{get_intervention_levels}
\title{Extract intervention levels from parameter names}
\usage{
get_intervention_levels(param_names)
}
\description{
Extract intervention levels from parameter names
}
