% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_risk.R
\name{tmle_risk}
\alias{tmle_risk}
\title{Risk Measures for Binary Outcomes}
\usage{
tmle_risk(baseline_level = NULL, effect_scale = "multiplicative")
}
\description{
Estimates TSMs, RRs, PAR, and PAF
}
\details{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome binary
}
