% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_risk_binary.R
\name{tmle_risk_binary}
\alias{tmle_risk_binary}
\title{Risk Measures for Binary Outcomes}
\usage{
tmle_risk_binary(baseline_level = NULL)
}
\description{
Estimates TSMs, RRs, PAR, and PAF
}
\details{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome binary
}
